Try to fix new warnings

+++ Modules/_ssl.c
@@ -2472,8 +2472,8 @@ _ssl__SSLSocket_sendfile_impl(PySSLSocket *self, int fd, Py_off_t offset,
 
     do {
         PySSL_BEGIN_ALLOW_THREADS
-        retval = SSL_sendfile(self->ssl, fd, offset, size, flags);
-        err = _PySSL_errno(retval < 0, self->ssl, retval);
+        retval = SSL_sendfile(self->ssl, fd, (off_t)offset, size, flags);
+        err = _PySSL_errno(retval < 0, self->ssl, (int)retval);
         PySSL_END_ALLOW_THREADS
         self->err = err;
 
@@ -2518,7 +2518,7 @@ _ssl__SSLSocket_sendfile_impl(PySSLSocket *self, int fd, Py_off_t offset,
     }
     Py_XDECREF(sock);
     if (retval < 0)
-        return PySSL_SetError(self, retval, __FILE__, __LINE__);
+        return PySSL_SetError(self, (int)retval, __FILE__, __LINE__);
     if (PySSL_ChainExceptions(self) < 0)
         return NULL;
     return PyLong_FromSize_t(retval);
